more work on DatabaseConnectionInfo
sensible reuse of the JDBC Connection
